change storage format of balance

+++ Core/Lyra.Core/Blocks/TransactionBlock.cs
@@ -33,7 +33,7 @@ namespace Lyra.Core.Blocks
         public string AccountID { get; set; }
 
         // this is the number of atomic units; it must be divided by the number of digits after the digital point for specific currency
-        public Dictionary<string, decimal> Balances { get; set; }
+        public Dictionary<string, long> Balances { get; set; }
         //public List<string, decimal> Balances { get; set; }
 
         public decimal Fee { get; set; } // the amount of transfer fee paid to the authorizers for processing transfer
